requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
pytz==2023.3
yarl==1.9.2
anyio==3.7.1
asgiref==3.7.2
Django==3.2.12
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
python-telegram-bot==20.4
sniffio==1.3.0
sqlparse==0.4.4